Optimize NuBase LaTeX string buffer
