+++ include/reactions/pdg.hpp.in
@@ -243,7 +243,7 @@ namespace reactions {
         // maximum size of an expansion of names to LaTeX
         // names.
         //
-        // Obtained with: scripts/pdg-latex-name-buffers.py
+        // Obtained with: scripts/latex-name-buffers.py
         //
         latex.reserve(4 * ref.size());
         superscript_buffer.reserve(
